jigsaw/prefect_flask_lab/prefect-flask-lab/codebase/requirements.txt
spotipy
python-dotenv
pandas
pytest
prefect
sqlalchemy
psycopg2-binary
flask